refactor circuit simulators with CircuitBase
